backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.0
google-generativeai>=0.8.0
Pillow>=10.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-multipart>=0.0.6
httpx>=0.25.0